requirements.txt
nltk
tensorflow-gpu==1.14.0
bert-tensorflow
tensorflow-hub
gensim
numpy
pandas
sklearn
matplotlib
ipython==7.7.0
prompt-toolkit==2.0.9